+++ spel_modul2/GameEngine/Systems/RenderSystems/RenderEnergySystem.cs
@@ -25,7 +25,7 @@ namespace GameEngine.Systems
 
                 if (energyComponent != null && healthComponent != null && healthComponent.IsAlive == true)
                 {
-                    int currEnergy = energyComponent.Current;
+                    float currEnergy = energyComponent.Current;
                     Rectangle energyRectangle = new Rectangle();
                     int playNum = cm.GetComponentForEntity<PlayerComponent>(entity.Key).Number;
                     float scaledEnergy = (float)currEnergy / energyComponent.Max * 100f;
